+++ typed-builder-macro/src/util.rs
@@ -1,10 +1,14 @@
+use std::{collections::HashSet, iter};
+
 use proc_macro2::{Ident, Span, TokenStream};
-use quote::ToTokens;
+use quote::{format_ident, ToTokens};
 use syn::{
     parenthesized,
-    parse::{Parse, Parser},
+    parse::{Parse, ParseStream, Parser},
+    parse_quote,
     punctuated::Punctuated,
-    token, Error, Expr, Token,
+    spanned::Spanned,
+    token, Attribute, Error, Expr, FnArg, ItemFn, Pat, PatIdent, ReturnType, Signature, Token, Type,
 };
 
 pub fn path_to_single_string(path: &syn::Path) -> Option<String> {
@@ -202,6 +206,16 @@ impl ToTokens for KeyValue {
     }
 }
 
+impl Parse for KeyValue {
+    fn parse(input: syn::parse::ParseStream) -> syn::Result<Self> {
+        Ok(Self {
+            name: input.parse()?,
+            eq: input.parse()?,
+            value: input.parse()?,
+        })
+    }
+}
+
 pub struct SubAttr {
     pub name: Ident,
     pub paren: token::Paren,
@@ -211,6 +225,9 @@ pub struct SubAttr {
 impl SubAttr {
     pub fn args<T: Parse>(self) -> syn::Result<impl IntoIterator<Item = T>> {
         Punctuated::<T, Token![,]>::parse_terminated.parse2(self.args)
+    }
+    pub fn undelimited<T: Parse>(self) -> syn::Result<impl IntoIterator<Item = T>> {
+        (|p: ParseStream| iter::from_fn(|| (!p.is_empty()).then(|| p.parse())).collect::<syn::Result<Vec<T>>>()).parse2(self.args)
     }
 }
 
@@ -288,5 +305,144 @@ pub trait ApplyMeta {
         }
 
         Ok(())
+    }
+
+    fn apply_attr(&mut self, attr: &Attribute) -> syn::Result<()> {
+        match &attr.meta {
+            syn::Meta::List(list) => self.apply_subsections(list),
+            meta => Err(Error::new_spanned(meta, "Expected builder(…)")),
+        }
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct Mutator {
+    pub fun: ItemFn,
+    pub required_fields: HashSet<Ident>,
+}
+
+#[derive(Default)]
+struct MutatorAttribute {
+    requires: HashSet<Ident>,
+}
+
+impl ApplyMeta for MutatorAttribute {
+    fn apply_meta(&mut self, expr: AttrArg) -> Result<(), Error> {
+        if expr.name() != "requires" {
+            return Err(Error::new_spanned(expr.name(), "Only `requires` is supported"));
+        }
+
+        match expr.key_value()?.value {
+            Expr::Array(syn::ExprArray { elems, .. }) => self.requires.extend(
+                elems
+                    .into_iter()
+                    .map(|expr| match expr {
+                        Expr::Path(path) if path.path.get_ident().is_some() => {
+                            Ok(path.path.get_ident().cloned().expect("should be ident"))
+                        }
+                        expr => Err(Error::new_spanned(expr, "Expected field name")),
+                    })
+                    .collect::<Result<Vec<_>, _>>()?,
+            ),
+            expr => {
+                return Err(Error::new_spanned(
+                    expr,
+                    "Only list of field names [field1, field2, …] supported",
+                ))
+            }
+        }
+        Ok(())
+    }
+}
+
+impl Parse for Mutator {
+    fn parse(input: ParseStream) -> syn::Result<Self> {
+        let mut fun: ItemFn = input.parse()?;
+
+        let mut attribute = MutatorAttribute::default();
+
+        let mut i = 0;
+        while i < fun.attrs.len() {
+            let attr = &fun.attrs[i];
+            if attr.path().is_ident("mutator") {
+                attribute.apply_attr(attr)?;
+                fun.attrs.remove(i);
+            } else {
+                i += 1;
+            }
+        }
+
+        // Ensure `&mut self` receiver
+        if let Some(FnArg::Receiver(receiver)) = fun.sig.inputs.first_mut() {
+            *receiver = parse_quote!(&mut self);
+        } else {
+            // Error either on first argument or `()`
+            return Err(syn::Error::new(
+                fun.sig
+                    .inputs
+                    .first()
+                    .map(Spanned::span)
+                    .unwrap_or(fun.sig.paren_token.span.span()),
+                "mutator needs to take a reference to `self`",
+            ));
+        };
+
+        Ok(Self {
+            fun,
+            required_fields: attribute.requires,
+        })
+    }
+}
+
+fn pat_to_ident(i: usize, pat: &Pat) -> Ident {
+    if let Pat::Ident(PatIdent { ident, .. }) = pat {
+        ident.clone()
+    } else {
+        format_ident!("__{i}", span = pat.span())
+    }
+}
+
+impl Mutator {
+    /// Signature for Builder::<mutator> function
+    pub fn outer_sig(&self, output: Type) -> Signature {
+        let mut sig = self.fun.sig.clone();
+        sig.output = ReturnType::Type(Default::default(), output.into());
+
+        sig.inputs = sig
+            .inputs
+            .into_iter()
+            .enumerate()
+            .map(|(i, input)| match input {
+                FnArg::Receiver(_) => parse_quote!(self),
+                FnArg::Typed(mut input) => {
+                    input.pat = Box::new(
+                        PatIdent {
+                            attrs: Vec::new(),
+                            by_ref: None,
+                            mutability: None,
+                            ident: pat_to_ident(i, &input.pat),
+                            subpat: None,
+                        }
+                        .into(),
+                    );
+                    FnArg::Typed(input)
+                }
+            })
+            .collect();
+        sig
+    }
+
+    /// Arguments to call inner mutator function
+    pub fn arguments(&self) -> Punctuated<Ident, Token![,]> {
+        self.fun
+            .sig
+            .inputs
+            .iter()
+            .enumerate()
+            .filter_map(|(i, input)| match &input {
+                FnArg::Receiver(_) => None,
+                FnArg::Typed(input) => Some(pat_to_ident(i, &input.pat)),
+            })
+            .collect()
     }
 }